Add support for running bolt without a bootstrap file.
This uses data attributes to control config location, main method and
whether define/require etc... should be registered as globals.

This is an optional mode, but it is friendly so you can set up
projects that can run from static files, without generated code
and without checking in bootstrap files.